# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Authentication",
        "description": "API de gestion des utilisateurs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/allergene/create": {
            "post": {
                "tags": [
                    "Allergène"
                ],
                "summary": "Créer un allergène",
                "description": "Crée un nouvel allergène. Réservé aux administrateurs et employés.",
                "operationId": "df8256787fd6c5fadf6ebb9378fe3077",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Gluten"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Allergène créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Allergène créé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/allergene/{id}": {
            "get": {
                "tags": [
                    "Allergène"
                ],
                "summary": "Afficher un allergène",
                "description": "Retourne les informations d'un allergène par son identifiant. Accessible publiquement.",
                "operationId": "d08c630ecae993fc82404625fb34374d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'allergène",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Allergène trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "libelle": {
                                            "type": "string",
                                            "example": "Arachide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Allergène introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Allergène"
                ],
                "summary": "Modifier un allergène",
                "description": "Met à jour le libellé d'un allergène existant. Réservé aux administrateurs et employés.",
                "operationId": "750da404931bf321f8837f97e024159e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'allergène",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Fruits à coque"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Allergène mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Allergène mis à jour avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "404": {
                        "description": "Allergène introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Allergène"
                ],
                "summary": "Supprimer un allergène",
                "description": "Supprime un allergène par son identifiant. Réservé aux administrateurs et employés.",
                "operationId": "e33a8629620078c0277d6b7c76b0f8b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'allergène",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Allergène supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Allergène supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "404": {
                        "description": "Allergène introuvable"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}